Removed duplicate of CogwheelBlockItem injector. Refactored chain conveyor handlers as they're not client-side. updated TODO list.
