+++ common/src/main/java/org/valkyrienskies/mod/mixin/mod_compat/create/TODO.md
@@ -6,8 +6,9 @@ MixinMinecartInstance - ~~commented out, og code moved to vanillin~~ This class 
 
 MixinFlwContraption - fix needs testing, ~~and transformLightboxToWorld commented out~~ LightBox was moved to ContraptionVisual. MixinContraptionVisual will handle it now. (by Bunting_chj)
 
-Elevator Contraption - Previous version had the elevator descend to abyss.
+Elevator Contraption - ~~Previous version had the elevator descend to abyss.~~ Not anymore so I think it's gone.
 
-Hose Pulley - Let's make them drain the world from the ship.
+Hose Pulley - ~~Let's make them drain the world from the ship.~~ I think this was PR'd by someone else.
 
-Package Entities - have problem on belts either accepting or ejecting them.
+Package Entities - ~~have problem on belts either accepting or ejecting them.~~ Solved!
+
