+++ common/src/main/resources/valkyrienskies-common.mixins.json
@@ -82,14 +82,14 @@
     "mod_compat.create.block.MixinRedstoneContactBlock",
     "mod_compat.create.block.MixinStickerBlock",
     "mod_compat.create.blockentity.IMixinMechanicalBearingTileEntity",
+    "mod_compat.create.blockentity.MixinChainConveyorInteractionHandler",
+    "mod_compat.create.blockentity.MixinChainConveyorRidingHandler",
+    "mod_compat.create.blockentity.MixinChainPackageInteractionHandler",
     "mod_compat.create.blockentity.MixinCrushingWheelControllerTileEntity",
     "mod_compat.create.blockentity.MixinEjectorTileEntity",
     "mod_compat.create.blockentity.MixinEncasedFanTileEntity",
     "mod_compat.create.blockentity.MixinSawBlockEntity",
     "mod_compat.create.blockentity.MixinSchematicannonBlockEntity",
-    "mod_compat.create.client.MixinChainConveyorInteractionHandler",
-    "mod_compat.create.client.MixinChainConveyorRidingHandler",
-    "mod_compat.create.client.MixinChainPackageInteractionHandler",
     "mod_compat.create.client.MixinContraptionVisual",
     "mod_compat.create.client.MixinDeployTool",
     "mod_compat.create.client.MixinFlwContraption",
@@ -177,7 +177,6 @@
     "mod_compat.create.client.MixinChainConveyorRenderer",
     "mod_compat.create.client.MixinChainConveyorVisual",
     "mod_compat.create.client.MixinCogwheelBlockItem",
-    "mod_compat.create.client.MixinCogwheelBlockItemHitOnShaft",
     "mod_compat.create.client.MixinContraptionControlsRenderer",
     "mod_compat.create.client.MixinContraptionRenderDispatcher",
     "mod_compat.create.client.MixinElevatorControlsHandler",
